requirements.txt
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".
fastapi
google-cloud-storage
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-sdk
opentelemetry-api
opentelemetry-exporter-gcp-monitoring
opentelemetry-exporter-gcp-trace
python-json-logger
python-json-logger
redis
uvicorn[standard]
